Fix seeding
The seed is expected to be a 64-bit unsigned integer, according to
the information returned from an unsuccessful API request.

This was working until recently. I don't know if the issue is that
the API has changed, or just that I was using it incorrectly before
but with less server-side validation.

The https://github.com/EliahKagan/algal repo was likewise changed:

https://github.com/EliahKagan/algal/commit/a9e3cb9bcf004d47f04e0bd05a6f08987a50d7fb